refactor: fix fetchers names by removing duplicating postfix (#4375)
