+++ .github/labeler.yml
@@ -28,7 +28,7 @@ lang-card:
       - any-glob-to-any-file:
           - api/top-langs.js
           - src/cards/top-languages-card.js
-          - src/fetchers/top-languages-fetcher.js
+          - src/fetchers/top-languages.js
           - tests/fetchTopLanguages.test.js
           - tests/renderTopLanguagesCard.test.js
           - tests/top-langs.test.js
@@ -38,7 +38,7 @@ repo-card:
       - any-glob-to-any-file:
           - api/pin.js
           - src/cards/repo-card.js
-          - src/fetchers/repo-fetcher.js
+          - src/fetchers/repo.js
           - tests/fetchRepo.test.js
           - tests/renderRepoCard.test.js
           - tests/pin.test.js
@@ -48,7 +48,7 @@ stats-card:
       - any-glob-to-any-file:
           - api/index.js
           - src/cards/stats-card.js
-          - src/fetchers/stats-fetcher.js
+          - src/fetchers/stats.js
           - tests/fetchStats.test.js
           - tests/renderStatsCard.test.js
           - tests/api.test.js
@@ -58,7 +58,7 @@ wakatime-card:
       - any-glob-to-any-file:
           - api/wakatime.js
           - src/cards/wakatime-card.js
-          - src/fetchers/wakatime-fetcher.js
+          - src/fetchers/wakatime.js
           - tests/fetchWakatime.test.js
           - tests/renderWakatimeCard.test.js
           - tests/wakatime.test.js
@@ -68,7 +68,7 @@ gist-card:
       - any-glob-to-any-file:
           - api/gist.js
           - src/cards/gist-card.js
-          - src/fetchers/gist-fetcher.js
+          - src/fetchers/gist.js
           - tests/fetchGist.test.js
           - tests/renderGistCard.test.js
           - tests/gist.test.js
