+++ api/gist.js
@@ -6,7 +6,7 @@ import {
 } from "../src/common/utils.js";
 import { isLocaleAvailable } from "../src/translations.js";
 import { renderGistCard } from "../src/cards/gist-card.js";
-import { fetchGist } from "../src/fetchers/gist-fetcher.js";
+import { fetchGist } from "../src/fetchers/gist.js";
 
 export default async (req, res) => {
   const {
