+++ tests/fetchRepo.test.js
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
-import { fetchRepo } from "../src/fetchers/repo-fetcher.js";
+import { fetchRepo } from "../src/fetchers/repo.js";
 import { expect, it, describe, afterEach } from "@jest/globals";
 
 const data_repo = {
