+++ tests/fetchStats.test.js
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
 import { calculateRank } from "../src/calculateRank.js";
-import { fetchStats } from "../src/fetchers/stats-fetcher.js";
+import { fetchStats } from "../src/fetchers/stats.js";
 import { expect, it, describe, beforeEach, afterEach } from "@jest/globals";
 
 // Test parameters.
